requirements.txt
strands-agents>=0.3.0
strands-agents-tools>=0.3.0
fastmcp>=2.14.0
mcp>=1.0.0
pandas>=2.1.0
streamlit>=1.32.0
python-dotenv>=1.0.1
boto3>=1.34.0
